requirements.txt
Pillow
matplotlib
datetime
chemicals
rdkit
mysql-connector-python